Replace `longjmp` with `throw`
When we use `longjmp`, non-trivial destructors of objects will not be called. This may become problematic in the future. We replaced it with a `throw` expression. Safety of using throw in a signal handler is not clear and is mostly compiler dependant. However, It seems that no bad things wil happen in most circumstances.

Signed-off-by: aybehrouz <behrouz_ayati@yahoo.com>