Fix some issues regarding argc strings
`STRING` macro was not allocating its memory on the stack. We know that allocating memory outside the stack could become problematic and result in nondeterministic behaviour of validators.

`response_buffer()` function returned a reference to an argc object, while returning argc objects is not allowed. We added response as a parameter to `dispatcher`'s signature to fix this issue.

Signed-off-by: aybehrouz <behrouz_ayati@yahoo.com>